+++ dataexplorer/global.R
@@ -50,7 +50,7 @@ globvars <- list(
   subsetVars = TRUE,
 
   # Nb max item when multiselect
-  nbopt_multiselect = 150,
+  nbopt_multiselect = 250,
 
   # Nb max variables for COR && GGM analyses
   max_multivars = 150,